Update script to handle how TR&Ds are changing.
Since TRD3 is becoming TRD4, and the new TRD3 was previously a part of
TRD2, we need to update how the tags are represented on the publication
page. The idea is to show what type of work (e.g. molecular modeling) it
actually belongs to instead of the changing and arbitrarily assigned TRD
number.